Move reference data components out of data-utils
