+++ src/data/reference-data/ReferenceDataDeleter.js
@@ -1,0 +1,34 @@
+import _ from 'lodash/fp';
+import { useState } from 'react';
+import { b, div, h } from 'react-hyperscript-helpers';
+import { absoluteSpinnerOverlay, DeleteConfirmationModal } from 'src/components/common';
+import { Ajax } from 'src/libs/ajax';
+import { reportError } from 'src/libs/error';
+
+import ReferenceData from './references';
+
+export const ReferenceDataDeleter = ({ onSuccess, onDismiss, namespace, name, referenceDataType }) => {
+  const [deleting, setDeleting] = useState(false);
+
+  return h(
+    DeleteConfirmationModal,
+    {
+      objectType: 'reference',
+      objectName: referenceDataType,
+      onConfirm: async () => {
+        setDeleting(true);
+        try {
+          await Ajax()
+            .Workspaces.workspace(namespace, name)
+            .deleteAttributes(_.map((key) => `referenceData_${referenceDataType}_${key}`, _.keys(ReferenceData[referenceDataType])));
+          onSuccess();
+        } catch (error) {
+          reportError('Error deleting reference data', error);
+          onDismiss();
+        }
+      },
+      onDismiss,
+    },
+    [div(['Are you sure you want to delete the ', b([referenceDataType]), ' reference data?']), deleting && absoluteSpinnerOverlay]
+  );
+};
